update API and add landing page
