+++ CC/.env
@@ -1,3 +1,3 @@
-GCS_BUCKET_NAME=tancap-bucket
+GCS_BUCKET_NAME="tancap-bucket"
 JWT_SECRET="E5uwQWuQRhLazoM295mA+8KJXsdFyGLBQceRbmA04kI="
 PORT=8080
