--- CC/Backend/app.yaml
@@ -1,10 +1,0 @@
-runtime: nodejs20 # or your Node.js version
-env_variables:
-
-  GCS_BUCKET_NAME: "tancap-bucket"
-  JWT_SECRET: "E5uwQWuQRhLazoM295mA+8KJXsdFyGLBQceRbmA04kI="
-  PORT: 8080
-
-handlers:
-- url: /.*
-  script: auto
