--- CC/Backend/routes/catalogRoutes.js
@@ -1,8 +1,0 @@
-module.exports = [
-    {
-      method: 'GET',
-      path: '/catalog',
-      handler: require('../handlers/catalogHandlers').getHandsigns
-    }
-  ];
-
