--- CC/Backend/routes/testRoutes.js
@@ -1,9 +1,0 @@
-// module.exports = [
-//     // ... (your other routes)
-//     {
-//       method: 'GET',
-//       path: '/firestore',  // Add a test endpoint
-//       handler: require('../handlers/testHandlers').testFirestoreConnection,
-//     },
-//   ];
-
