+++ CC/routes/catalogRoutes.js
@@ -1,0 +1,7 @@
+module.exports = [
+  {
+    method: 'GET',
+    path: '/api/catalog', 
+    handler: require('../handlers/catalogHandlers').getHandsigns
+  }
+];
