+++ CC/src/index.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Tancap - Empowering Sign Language</title>
+    <link rel="stylesheet" href="styles/index.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:wght@400;500;700&display=swap" rel="stylesheet">
+</head>
+<body>
+    <header>
+        <div class="logo">
+            <img src="assets/logo.png" alt="Tancap Logo">
+            <span>Tancap</span>
+        </div>
+        <nav>
+            <ul>
+                <li><a href="#">Home</a></li>
+                <li><a href="#">About</a></li>
+                <li><a href="#">News</a></li>
+            </ul>
+        </nav>
+        <div class="download-button">
+            <a href="#">Download</a>
+        </div>
+    </header>
+
+    <main>
+        <section class="hero">
+            <h1>Empowering <span class="highlight-1">Sign </span> <span class="highlight-2">Language</span> with Instant Translation</h1>
+            <img class="illustrations" src="assets/illust.png" alt="">
+        </section>
+
+        <section class="features">
+            <h2>Communication Shouldn't Be a Barrier</h2>
+            <div class="feature-cards">
+                <div class="card">
+                    <img src="assets/realtime-detection.png" alt="Real-Time Detection">
+                    <h3>Real-Time Detection</h3>
+                    <p>Tancap recognizes and translates your handsigns into text</p>
+                </div>
+                <div class="card">
+                    <img src="assets/stay-informed.png" alt="Stay Informed">
+                    <h3>Stay Informed</h3>
+                    <p>Access news and stories relevant to the deaf and mute community</p>
+                </div>
+                <div class="card">
+                    <img src="assets/expand your vocabulary.png" alt="Expand Your Vocabulary">
+                    <h3>Expand Your Vocabulary</h3>
+                    <p>Explore our comprehensive catalogue of handsigns and their meanings</p>
+                </div>
+            </div>
+        </section>
+    </main>
+
+    <footer>
+        <!-- Add footer content if needed -->
+    </footer>
+</body>
+</html>
