+++ CC/src/styles/index.css
@@ -1,0 +1,132 @@
+body {
+    font-family: 'Plus Jakarta Sans', sans-serif;
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+}
+
+header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 20px 4rem;
+    background-color: white;
+}
+
+header .logo {
+    display: flex;
+    align-items: center;
+    font-weight: 700;
+    font-size: 2rem;
+}
+
+header .logo img {
+    width: 60px;
+    margin-right: 10px;
+}
+
+header .logo span {
+    background: linear-gradient(180deg, #8DC165 0%, #3CB8B6 100%);
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
+
+header nav ul {
+    list-style: none;
+    display: flex;
+    gap: 6rem;
+}
+
+header nav ul li {
+    display: inline;
+}
+
+header nav ul li a {
+    text-decoration: none;
+    color: black;
+    font-weight: 500;
+    font-size: 1.75rem;
+}
+
+header .download-button a {
+    text-decoration: none;
+    padding: 10px 2rem;
+    background-color: #00C6D7;
+    color: white;
+    font-weight: 700;
+    font-size: 1.5rem;
+    border-radius: 25px;
+}
+
+.hero {
+    display: flex;
+    justify-content: space-around;
+    text-align: left;
+    padding: 50px 20px;
+    background-color: #f3f3f3;
+}
+
+.hero h1 {
+    max-width: 500px;
+    font-size: 4rem;
+    margin: 0;
+    text-align: left;
+}
+
+.hero .highlight-1 {
+    color: #8DC165;
+}
+
+.hero .highlight-2 {
+    color: #3CB8B6;
+}
+
+.hero .illustrations {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+    height: 25rem;
+}
+
+.features {
+    background-color: #007B80;
+    color: white;
+    text-align: center;
+    padding: 20px 4rem;
+}
+
+.features h2 {
+    margin-bottom: 40px;
+    font-size: 3rem;
+}
+
+.features .feature-cards {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+}
+
+.features .card {
+    background-color: rgba(255, 255, 255, 0);
+    color: rgb(255, 255, 255);
+    padding: 20px;
+    border-radius: 10px;
+    width: 650px;
+    text-align: center;
+}
+
+.features .card img {
+    width: 100%;
+    height: auto;
+}
+
+.features .card h3 {
+    margin: 20px 0 15px;
+    font-size: 2rem;
+}
+
+.features .card p {
+    margin: 0;
+    font-size: 1.5rem;
+    margin-bottom: 2rem;
+}
